10/31 村人の通常単体攻撃、雑魚敵の範囲攻撃、雑魚敵が村人に接するまで移動する（ランダムにストップする要素も入れる）

+++ Assets/code/Attack_Object.cs
@@ -6,9 +6,9 @@ using UnityEngine;
 
 public class Attack_Object : MonoBehaviour
 {
-    public float attack_point;//UŒ‚—Í
+    public int attack_point;//UŒ‚—Í
     public List<string> features_point;//‚©‚©‚è‚â‚·‚¢ó‘Ô
-    public void Initialize(float c_attack_point, List<string> c_features_point)
+    public void Initialize(int c_attack_point, List<string> c_features_point)
     {
         attack_point = c_attack_point;
         features_point = c_features_point;
